+++ src/Farmer/Builders/Builders.Cdn.fs
@@ -17,7 +17,7 @@ type EndpointConfig =
       Http : FeatureFlag
       Https : FeatureFlag
       Compression : FeatureFlag
-      Origin : string
+      Origin : ArmExpression
       CustomDomain : Uri option
       OptimizationType : OptimizationType }
 
@@ -25,7 +25,7 @@ type CdnConfig =
     { Name : ResourceName
       Sku : Sku
       Endpoints : EndpointConfig list
-      Tags: Map<string,string>  }
+      Tags: Map<string,string> }
     interface IBuilder with
         member this.DependencyName = this.Name
         member this.BuildResources _ = [
@@ -47,6 +47,7 @@ type CdnConfig =
                 match endpoint.CustomDomain with
                 | Some customDomain ->
                     { Name = endpoint.Name.Map(sprintf "%sdomain")
+                      Profile = this.Name
                       Endpoint = endpoint.Name
                       Hostname = customDomain }
                 | None ->
@@ -81,7 +82,7 @@ type EndpointBuilder() =
           Http = Enabled
           Https = Enabled
           Compression = Disabled
-          Origin = ""
+          Origin = ArmExpression.Empty
           CustomDomain = None
           OptimizationType = GeneralWebDelivery }
 
@@ -91,10 +92,12 @@ type EndpointBuilder() =
     member this.Name(state:EndpointConfig, name) = this.Name(state, ResourceName name)
     /// The address of the origin.
     [<CustomOperation "origin">]
-    member _.Origin(state:EndpointConfig, name) =
+    member _.Origin(state:EndpointConfig, name:ArmExpression) =
       { state with
-          Name = state.Name.IfEmpty ((name |> Seq.filter Char.IsLetterOrDigit |> Seq.toArray |> String) + "-endpoint")
+          Name = state.Name.IfEmpty ((name.Value |> Seq.filter Char.IsLetterOrDigit |> Seq.toArray |> String) + "-endpoint")
           Origin = name }
+    member this.Origin(state:EndpointConfig, name) = this.Origin(state, ArmExpression.literal name)
+    member this.Origin(state:EndpointConfig, name:Uri) = this.Origin(state, ArmExpression.literal name.Host)
 
     member private _.AddDependency (state:EndpointConfig, resourceName:ResourceName) = { state with Dependencies = ResourceId.create resourceName :: state.Dependencies }
     member private _.AddDependencies (state:EndpointConfig, resourceNames:ResourceName list) = { state with Dependencies = (resourceNames |> List.map ResourceId.create) @ state.Dependencies }